Trait acpi_tables::aml::Aml

pub trait Aml {
    fn to_aml_bytes(&self, bytes: &mut Vec<u8>);
}

The trait Aml can be implemented by the ACPI objects to translate itself into the AML raw data. So that these AML raw data can be added into the ACPI DSDT for guest.

Required methods

fn to_aml_bytes(&self, bytes: &mut Vec<u8>)

Translate an ACPI object into AML code and append to the vector buffer.

bytes - The vector used to append the AML code.